// src/app/swaggerFiles/model/adulte.ts
/**
 * assoparenteleve API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Eleve } from './eleve';
import { Role } from './role';

export interface Adulte { 
    id?: number;
    nom?: string;
    prenom?: string;
    mail?: string;
    role?: Role;
    motDePasse?: string;
    telephone?: string;
    eleves?: Array<Eleve>;
    active?: boolean;
}